+++ allel/abc.py
@@ -2,7 +2,119 @@
 from __future__ import absolute_import, print_function, division
 
 
-from .abc import ArrayWrapper
+# third-party imports
+import numpy as np
+
+
+class ArrayWrapper(object):
+    """Abstract base class that delegates everything to a wrapped array-like object."""
+
+    def __init__(self, values):
+        self._values = values
+
+    @property
+    def values(self):
+        """The underlying array of values."""
+        return self._values
+
+    def __getattr__(self, item):
+        return getattr(self.values, item)
+
+    def __getitem__(self, item):
+        return self.values[item]
+
+    def __setitem__(self, item, value):
+        self.values[item] = value
+
+    def __iter__(self):
+        return iter(self.values)
+
+    def __len__(self):
+        return len(self.values)
+
+    def __array__(self, *args):
+        a = np.asarray(self.values)
+        if args:
+            a = a.astype(args[0])
+        return a
+
+    def __str__(self):
+        return str(self.values)
+
+    def __repr__(self):
+        return repr(self.values)
+
+    def __eq__(self, other):
+        return self.values == other
+
+    def __ne__(self, other):
+        return self.values != other
+
+    def __lt__(self, other):
+        return self.values < other
+
+    def __gt__(self, other):
+        return self.values > other
+
+    def __le__(self, other):
+        return self.values <= other
+
+    def __ge__(self, other):
+        return self.values >= other
+
+    def __abs__(self):
+        return abs(self.values)
+
+    def __add__(self, other):
+        return self.values + other
+
+    def __and__(self, other):
+        return self.values & other
+
+    def __div__(self, other):
+        return self.values.__div__(other)
+
+    def __floordiv__(self, other):
+        return self.values // other
+
+    def __inv__(self):
+        return ~self.values
+
+    def __invert__(self):
+        return ~self.values
+
+    def __lshift__(self, other):
+        return self.values << other
+
+    def __mod__(self, other):
+        return self.values % other
+
+    def __mul__(self, other):
+        return self.values * other
+
+    def __neg__(self):
+        return -self.values
+
+    def __or__(self, other):
+        return self.values | other
+
+    def __pos__(self):
+        return +self.values
+
+    def __pow__(self, other):
+        return self.values ** other
+
+    def __rshift__(self, other):
+        return self.values >> other
+
+    def __sub__(self, other):
+        return self.values - other
+
+    def __truediv__(self, other):
+        return self.values.__truediv__(other)
+
+    def __xor__(self, other):
+        return self.values ^ other
 
 
 def arr1d_to_html(indices, items, caption):
@@ -11,7 +123,7 @@ def arr1d_to_html(indices, items, caption):
 
     # sanitize caption
     caption = caption.replace('<', '&lt;')
-    caption = caption.strip().replace('\n', '<br/>')
+    # caption = caption.strip().replace('\n', '<br/>')
     html = caption
 
     # build table
@@ -35,7 +147,7 @@ def arr2d_to_html(row_indices, col_indices, items, caption):
 
     # sanitize caption
     caption = caption.replace('<', '&lt;')
-    caption = caption.strip().replace('\n', '<br/>')
+    # caption = caption.strip().replace('\n', '<br/>')
     html = caption
 
     # build table
@@ -65,7 +177,7 @@ def recarr_to_html(names, indices, items, caption):
 
     # sanitize caption
     caption = caption.replace('<', '&lt;')
-    caption = caption.strip().replace('\n', '<br/>')
+    # caption = caption.strip().replace('\n', '<br/>')
     html = caption
 
     # build table
@@ -215,7 +327,7 @@ class Displayable2D(Displayable):
 
     def to_display_html(self, row_threshold=6, col_threshold=10, row_edgeitems=3,
                         col_edgeitems=5, caption=None):
-        row_indices, col_indices, items = self._display_items(
+        row_indices, col_indices, items = self.get_display_items(
             row_threshold, col_threshold, row_edgeitems, col_edgeitems
         )
         if caption is None:
