+++ allel/model/__init__.py
@@ -8,7 +8,7 @@ from .ndarray import *
 from . import chunked
 from .chunked import *
 from . import util
-from .util import create_allele_mapping, locate_fixed_differences, locate_private_alleles
+from .util import *
 
 # from allel.model import chunked
